# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JffEhr.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/CrewMedicalCheckins": {
      "get": {
        "tags": [
          "CrewMedicalCheckins"
        ],
        "parameters": [
          {
            "name": "campId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "crewId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CrewMedicalCheckins"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCrewMedicalCheckinRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCrewMedicalCheckinRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCrewMedicalCheckinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/CrewMedicalCheckins/{id}": {
      "get": {
        "tags": [
          "CrewMedicalCheckins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrewMedicalCheckinDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CrewMedicalCheckins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCrewMedicalCheckinRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCrewMedicalCheckinRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCrewMedicalCheckinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CrewMedicalCheckins"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/EmergencyContacts": {
      "get": {
        "tags": [
          "EmergencyContacts"
        ],
        "parameters": [
          {
            "name": "camperId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmergencyContactDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmergencyContactDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmergencyContactDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EmergencyContacts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmergencyContactRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmergencyContactRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmergencyContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/EmergencyContacts/{id}": {
      "get": {
        "tags": [
          "EmergencyContacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EmergencyContacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmergencyContactRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmergencyContactRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmergencyContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "EmergencyContacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCrewMedicalCheckinRequest": {
        "required": [
          "campId",
          "crewId"
        ],
        "type": "object",
        "properties": {
          "crewId": {
            "type": "string",
            "format": "uuid"
          },
          "campId": {
            "type": "string",
            "format": "uuid"
          },
          "allergies": {
            "type": "string",
            "nullable": true
          },
          "hasBroviacPort": {
            "type": "boolean"
          },
          "eyesight": {
            "type": "string",
            "nullable": true
          },
          "hearing": {
            "type": "string",
            "nullable": true
          },
          "mobilityAids": {
            "type": "string",
            "nullable": true
          },
          "currentMedications": {
            "type": "string",
            "nullable": true
          },
          "medicalReleaseSigned": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CreateEmergencyContactRequest": {
        "required": [
          "camperId",
          "cellNo",
          "name",
          "relationship"
        ],
        "type": "object",
        "properties": {
          "camperId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "cellNo": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          },
          "workNo": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "relationship": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "CrewMedicalCheckinDto": {
        "type": "object",
        "properties": {
          "checkinId": {
            "type": "string",
            "format": "uuid"
          },
          "crewId": {
            "type": "string",
            "format": "uuid"
          },
          "crewName": {
            "type": "string",
            "nullable": true
          },
          "campId": {
            "type": "string",
            "format": "uuid"
          },
          "allergies": {
            "type": "string",
            "nullable": true
          },
          "hasBroviacPort": {
            "type": "boolean"
          },
          "eyesight": {
            "type": "string",
            "nullable": true
          },
          "hearing": {
            "type": "string",
            "nullable": true
          },
          "mobilityAids": {
            "type": "string",
            "nullable": true
          },
          "currentMedications": {
            "type": "string",
            "nullable": true
          },
          "medicalReleaseSigned": {
            "type": "boolean"
          },
          "checkedInBy": {
            "type": "string",
            "format": "uuid"
          },
          "checkedInByName": {
            "type": "string",
            "nullable": true
          },
          "checkedInAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "EmergencyContactDto": {
        "required": [
          "cellNo",
          "name",
          "relationship"
        ],
        "type": "object",
        "properties": {
          "contactId": {
            "type": "string",
            "format": "uuid"
          },
          "camperId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cellNo": {
            "type": "string",
            "nullable": true
          },
          "workNo": {
            "type": "string",
            "nullable": true
          },
          "relationship": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCrewMedicalCheckinRequest": {
        "type": "object",
        "properties": {
          "allergies": {
            "type": "string",
            "nullable": true
          },
          "hasBroviacPort": {
            "type": "boolean"
          },
          "eyesight": {
            "type": "string",
            "nullable": true
          },
          "hearing": {
            "type": "string",
            "nullable": true
          },
          "mobilityAids": {
            "type": "string",
            "nullable": true
          },
          "currentMedications": {
            "type": "string",
            "nullable": true
          },
          "medicalReleaseSigned": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateEmergencyContactRequest": {
        "required": [
          "cellNo",
          "name",
          "relationship"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "cellNo": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string"
          },
          "workNo": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "relationship": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}